Refactored formEditor layout and added page context in it
