--- src/app/(formEditor)/layout.tsx
@@ -1,7 +1,0 @@
-type Props = {
-  children: React.ReactNode;
-};
-
-export default async function AuthLayout({ children }: Readonly<Props>) {
-  return <div>{children}</div>;
-}
