+++ src/app/(formEditorPage)/forms/[formId]/page.tsx
@@ -1,0 +1,5 @@
+import FormPreview from "@/components/formEditor/formPreview";
+
+export default async function NewFormPage() {
+  return <FormPreview />;
+}
