+++ src/components/copyLinkButton.tsx
@@ -1,0 +1,25 @@
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "./ui/tooltip";
+import { Copy } from "lucide-react";
+import { Button } from "./ui/button";
+
+export const CopyLinkButton = ({ onClick }: { onClick: () => void }) => {
+  return (
+    <TooltipProvider delayDuration={200}>
+      <Tooltip>
+        <TooltipTrigger asChild>
+          <Button variant="ghost" size="icon" onClick={onClick}>
+            <Copy className="h-4 w-4" />
+          </Button>
+        </TooltipTrigger>
+        <TooltipContent>
+          <p>Copy link</p>
+        </TooltipContent>
+      </Tooltip>
+    </TooltipProvider>
+  );
+};
